ddr_pipeline/requirements.txt
PyMuPDF>=1.24.0
Pillow>=10.0.0
google-generativeai>=0.5.0
python-docx>=1.1.0
python-dotenv>=1.0.0
